Article Description - class renaming & cleanup
WikidataDescriptionController > WikidataFetcher
WikidataFetcher > WikidataFetcher+Places
Removing wikidataDescriptionEditing reference from MWKDataStore
