+++ WMF Framework/Remote Notifications/RemoteNotificationsOperationsController.swift
@@ -32,7 +32,7 @@ class RemoteNotificationsOperationsController: NSObject {
         
         super.init()
 
-        NotificationCenter.default.addObserver(self, selector: #selector(didMakeAuthorizedWikidataDescriptionEdit), name: WikidataDescriptionEditingController.DidMakeAuthorizedWikidataDescriptionEditNotification, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(didMakeAuthorizedWikidataDescriptionEdit), name: WikidataFetcher.DidMakeAuthorizedWikidataDescriptionEditNotification, object: nil)
         NotificationCenter.default.addObserver(self, selector: #selector(modelControllerDidLoadPersistentStores(_:)), name: RemoteNotificationsModelController.didLoadPersistentStoresNotification, object: nil)
     }
