+++ Wikipedia/Code/WikidataFetcher.swift
@@ -1,66 +1,112 @@
-import Foundation
-import WMF
-import MapKit
+public struct WikidataAPIResult: Decodable {
+    public struct APIError: Error, Decodable {
+        public let code, info: String?
 
-class WikidataFetcher: Fetcher {
-    func wikidata(forArticleURL articleURL: URL, failure: @escaping (Error) -> Void, success: @escaping ([String: Any]) -> Void) {
-        guard let title = articleURL.wmf_title,
-            let language = articleURL.wmf_language else {
-                failure(RequestError.invalidParameters)
-                return
-        }
-        
-        let queryParameters = ["action": "wbgetentities", "title": title, "sites": "\(language)wiki", "format": "json"]
-        let components = configuration.wikidataAPIURLComponents(with: queryParameters)
-        session.getJSONDictionary(from: components.url, ignoreCache: false) { (jsonObject, response, error) in
-            guard let jsonObject = jsonObject else {
-                failure(RequestError.unexpectedResponse)
-                return
-            }
-            success(jsonObject)
+        public var localizedDescription: String {
+            return info ?? CommonStrings.unknownError
         }
     }
+    let error: APIError?
+    let success: Int?
+}
+
+struct MediaWikiSiteInfoResult: Decodable {
+    struct MediaWikiQueryResult: Decodable {
+        struct MediaWikiGeneralResult: Decodable {
+            let lang: String
+        }
+        let general: MediaWikiGeneralResult
+    }
+    let query: MediaWikiQueryResult
+}
+
+extension WikidataAPIResult {
+    var succeeded: Bool {
+        return success == 1
+    }
+}
+
+enum WikidataPublishingError: LocalizedError {
+    case invalidArticleURL
+    case apiResultNotParsedCorrectly
+    case notEditable
+    case unknown
+}
+
+public enum ArticleDescriptionSource: String {
+    case none
+    case unknown
+    case central
+    case local
     
-    func wikidataBoundingRegion(forArticleURL articleURL: URL, failure: @escaping (Error) -> Void, success: @escaping (MKCoordinateRegion) -> Void) {
-        wikidata(forArticleURL: articleURL, failure: failure) { (jsonObject) in
-            guard let entities = jsonObject["entities"] as? [String: Any],
-                let entity = entities.values.first as? [String: Any],
-                let claims = entity["claims"] as? [String: Any] else {
-                failure(RequestError.unexpectedResponse)
-                return
-            }
-            
-            let keys = ["P1332", "P1333", "P1334", "P1335"] //bounding coordinates
+    public static func from(string: String?) -> ArticleDescriptionSource {
+        guard let sourceString = string else {
+            return .none
+        }
+        guard let source = ArticleDescriptionSource(rawValue: sourceString) else {
+            return .unknown
+        }
+        return source
+    }
+}
 
-            let coordinates = keys.compactMap({ (key) -> CLLocationCoordinate2D? in
-                guard let values = claims[key] as? [Any] else {
-                    return nil
-                }
-                guard let point = values.first as? [String: Any] else {
-                    return nil
-                }
-                guard let mainsnak = point["mainsnak"] as? [String: Any] else {
-                    return nil
-                }
-                guard let datavalue = mainsnak["datavalue"] as? [String: Any] else {
-                    return nil
-                }
-                guard let value = datavalue["value"] as? [String: Any] else {
-                    return nil
-                }
-                guard let latitude = value["latitude"] as? CLLocationDegrees,
-                    let longitude = value["longitude"] as? CLLocationDegrees else {
-                        return nil
-                }
-                return CLLocationCoordinate2D(latitude: latitude, longitude: longitude)
-            })
-            
-            guard coordinates.count > 3 else {
-                failure(RequestError.unexpectedResponse)
+@objc public final class WikidataFetcher: Fetcher {
+    static let DidMakeAuthorizedWikidataDescriptionEditNotification = NSNotification.Name(rawValue: "WMFDidMakeAuthorizedWikidataDescriptionEdit")
+    /// Publish new wikidata description.
+    ///
+    /// - Parameters:
+    ///   - newWikidataDescription: new wikidata description to be published, e.g., "Capital of England and the United Kingdom".
+    ///   - source: description source; none, central or local.
+    ///   - wikidataID: id for the Wikidata entity including the prefix
+    ///   - language: language code of the page's wiki, e.g., "en".
+    ///   - completion: completion block called when operation is completed.
+    public func publish(newWikidataDescription: String, from source: ArticleDescriptionSource, forWikidataID wikidataID: String, language: String, completion: @escaping (Error?) -> Void) {
+        guard source != .local else {
+            completion(WikidataPublishingError.notEditable)
+            return
+        }
+        let requestWithCSRFCompletion: (WikidataAPIResult?, URLResponse?, Bool?, Error?) -> Void = { result, response, authorized, error in
+            if let error = error {
+                completion(error)
+            }
+            guard let result = result else {
+                completion(WikidataPublishingError.apiResultNotParsedCorrectly)
                 return
             }
 
-            success(coordinates.wmf_boundingRegion(with: 1))
+            completion(result.error)
+
+            if let authorized = authorized, authorized, result.error == nil {
+                DispatchQueue.main.async {
+                    NotificationCenter.default.post(name: WikidataFetcher.DidMakeAuthorizedWikidataDescriptionEditNotification, object: nil)
+                }
+            }
+        }
+        
+        let languageCodeParameters = WikipediaSiteInfo.defaultRequestParameters
+
+        let languageCodeComponents = configuration.mediaWikiAPIURLForWikiLanguage(language, with: languageCodeParameters)
+        session.jsonDecodableTask(with: languageCodeComponents.url) { (siteInfo: MediaWikiSiteInfoResult?, response, error) in
+            let normalizedLanguage = siteInfo?.query.general.lang ?? "en"
+            let queryParameters = ["action": "wbsetdescription",
+                                   "format": "json",
+                                   "formatversion": "2"]
+            let components = self.configuration.wikidataAPIURLComponents(with: queryParameters)
+            self.requestMediaWikiAPIAuthToken(for: components.url, type: .csrf) { (result) in
+                switch result {
+                case .failure(let error):
+                    completion(error)
+                case .success(let token):
+                    let bodyParameters = ["language": normalizedLanguage,
+                                          "uselang": normalizedLanguage,
+                                          "id": wikidataID,
+                                          "value": newWikidataDescription,
+                                          "token": token.value]
+                    self.session.jsonDecodableTask(with: components.url, method: .post, bodyParameters: bodyParameters, bodyEncoding: .form) { (result: WikidataAPIResult?, response, error) in
+                        requestWithCSRFCompletion(result, response, token.isAuthorized, error)
+                    }
+                }
+            }
         }
     }
 }
